Refactors dashboard layout and improves navigation
Replaces the existing dashboard layout with a more streamlined ClientDashboard component.

Updates the My Clothes page to indicate it is under construction, enhancing user experience.

Implements a responsive sidebar with navigation links, improving accessibility and user interaction.

Enhances the NavBar to dynamically display page titles based on the current path.

Improves the overall structure and organization of the dashboard for better maintainability.